File creation, execution, listing, editing
It is now possible to create files, edit them, and execute them.
However, files can only be a single block long and there is no way to
extend files to multiple blocks, which is the next thing to add. We need
to be able to append and prepend to files.