Zero length file creation, block rounding
Files are created with zero bytes in them using `mkfile` / `touch`.
`fallocate` can also handle `0` as an argument. Zero length files are
still backed by a single block (and thus 1024 can be used in that file
before another block has to be allocated for it).

When editing files with the block editor file lengths are automatically
rounded up to the next 1024 byte boundary.