Sped up `df`, implemented multiple FAT blocks
* Multiple FAT blocks have been added, or are allowed to be used, some
of the documentation needs to be updated to reflect this. This allows up
to 64MiB file systems to be created (tens of kilobytes less than this
really), thus removing a big restriction on he file system.
* Added `eof?-file` and `error?-file` to check if files are at the end
or in an error state.
* The `df` command has been greatly sped up, however it is still very
slow when used with large file systems, the culprit is the word
`largest`, which finds the largest allocatable block.