Commented out `resolve` code updated
Although still not working, the code for `resolve` has been updated.
This should parse out file names and return information about those file
names (such as whether the file exists, or it is actually a directory,
and more). This should eventually replace the file handling in each of
the commands, which apart from a few limited cases only handle directory
entries in the Present Working Directory.